# api/types/LoginResp.go
package types

import (
	"test_task/domain"
	"time"
)

type LoginResp struct {
	SessionId             string      `json:"sessionId"`
	AccessToken           string      `json:"access_token"`
	RefreshToken          string      `json:"refresh_token"`
	AccessTokenExpiresAt  time.Time   `json:"access_token_expires_at"`
	RefreshTokenExpiresAt time.Time   `json:"refresh_token_expires_at"`
	User                  domain.User `json:"user"`
}